[A3] Phase 32 Post-Completion: Code Simplifications & Maintenance
Refactor reduce crate for improved code quality:
- Replace manual Default trait implementations with #[default] derive
  - adaptive_classifier: AccessPatternType, CompressionLevel, DedupStrength, S3TieringPolicy
  - similarity_tier_stats: TierStats
- Use saturating_sub for overflow-safe arithmetic (similarity_tier_stats)
- Clean up boilerplate: -31 LOC net, zero functional changes
- All 2,071 unit tests pass ✅

This is part of ongoing maintenance following Phase 32 completion.

+++ crates/claudefs-reduce/src/dedup_bloom.rs
@@ -18,12 +18,12 @@ impl Default for BloomConfig {
 impl BloomConfig {
     pub fn bit_count(&self) -> usize {
         let n = self.expected_items as f64;
-        let p = self.false_positive_rate as f64;
+        let p = self.false_positive_rate;
         (n * (-1.44) * p.ln() / std::f64::consts::LN_2 / std::f64::consts::LN_2).ceil() as usize
     }
 
     pub fn hash_count(&self) -> usize {
-        ((-(self.false_positive_rate as f64).ln() / std::f64::consts::LN_2).ceil() as usize).max(1)
+        ((-self.false_positive_rate.ln() / std::f64::consts::LN_2).ceil() as usize).max(1)
     }
 }
